Migrated code to PSR-4
