--- layers/Schema/packages/migrate-users-wp/migrate/plugins/load.php
@@ -1,5 +1,0 @@
-<?php
-
-if (class_exists('\PoPSchema\CustomPostsWP\Component')) {
-	require_once 'pop-customposts-wp/load.php';
-}
