--- layers/Schema/packages/migrate-users-wp/migrate/plugins/pop-customposts-wp/kernel/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'cms-functionapi-hooks.php';
